+++ src/ua/data_types/node_attributes/method_attributes.rs
@@ -1,0 +1,13 @@
+impl super::MethodAttributes {
+    #[must_use]
+    pub fn with_executable(mut self, executable: bool) -> Self {
+        self.0.executable = executable;
+        self
+    }
+
+    #[must_use]
+    pub fn with_user_executable(mut self, user_executable: bool) -> Self {
+        self.0.userExecutable = user_executable;
+        self
+    }
+}